fovs added to project

+++ src/main/java/ar/edu/itba/it/cg/yart/utils/MathOperations.java
@@ -2,7 +2,7 @@ package ar.edu.itba.it.cg.yart.utils;
 
 public final class MathOperations {
 
-	public static int lerp(final int a, final int b, final int t) {
+	public static double lerp(final double a, final double b, final double t) {
 		return (1 - t) * a + (t * b);
 	}
